Added very slow mode for shooter
Since the speed we need for demos is slower than the tach can go, I
added a method to the shooter class that lets the speed be controlled
manually, which disables the closed-loop controller, and changed
robot3238.cpp to shoot at 25% power at the slowest speed, 3300RPM mid
and full blast at the highest setting. This should let us do demos
pretty well.